--- a/source/partials/application/_imported_docs.html.erb
+++ b/source/partials/application/_imported_docs.html.erb
@@ -1,24 +1,30 @@
 <% list_items = capture do %>
-  <% nested_docs.each do |page| %>
-    <% if page[:path] %>
-      <%= "<li><a href='#{page[:path]}' class='govuk-link'>#{page[:title]}</a></li>" %>
+  <%
+    docs_grouped_by_parent = imported_docs.each_with_object({}) do |page, hash|
+      path_to_dir = page[:path].split("/").take(page[:path].split("/").count - 1).join("/")
+      hash[path_to_dir] = [] unless hash[path_to_dir]
+      hash[path_to_dir] << page
+      hash[path_to_dir].sort_by! { |hash| hash[:title] }
+    end
+  %>
+  <% docs_grouped_by_parent.keys.each do |subpath| %>
+    <% list_of_links = capture do %>
+      <% docs_grouped_by_parent[subpath].each do |page| %>
+        <%= "<li><a href='#{page[:path]}' class='govuk-link'>#{page[:title]}</a></li>" %>
+      <% end %>
+    <% end %>
+
+    <% if subpath == path_to_app_docs %>
+      <%= list_of_links %>
     <% else %>
-      <li><strong><%= page[:title] %></strong></li>
-      <%= partial("partials/application/imported_docs", locals: { nested_docs: page[:docs] }).strip %>
+      <li><strong><%= subpath.sub(path_to_app_docs, "") %></strong></li>
+      <ul>
+        <%= list_of_links %>
+      </ul>
     <% end %>
   <% end %>
 <% end %>
 
-<%#
-  This is a hack to ensure markdown headings below the table continue to render properly.
-  See:
-  https://github.com/alphagov/govuk-developer-docs/pull/2519/commits/81726d015e43f8118d2f80b409d2a83325fbc803
-%>
-<% uls_on_newlines ||= false %>
-<% if uls_on_newlines %>
 <ul>
 <%= list_items.strip %>
 </ul>
-<% else %>
-<ul><%= list_items.strip %></ul>
-<% end %>