--- a/source/stylesheets/screen.css.scss
+++ b/source/stylesheets/screen.css.scss
@@ -1,5 +1,6 @@
 @import "govuk_tech_docs";
 @import "govuk_publishing_components/components/_summary-list";
+@import "govuk_publishing_components/components/_breadcrumbs";
 
 .gem-c-summary-list {
   border-bottom: 0; /* We have an extra border-bottom for some reason */