--- a/source/templates/application_template.html.md.erb
+++ b/source/templates/application_template.html.md.erb
@@ -58,8 +58,8 @@ source_url: https://github.com/alphagov/govuk-developer-docs/blob/master/data/ap
 <%= partial(
   "partials/application/imported_docs",
   locals: {
-    nested_docs: GitHubRepoFetcher.instance.nest_docs(imported_docs),
-    uls_on_newlines: true,
+    path_to_app_docs: "/apps/#{application.github_repo_name}",
+    imported_docs: imported_docs,
   }).strip %>
 <%
     end