--- a/source/templates/application_template.html.md.erb
+++ b/source/templates/application_template.html.md.erb
@@ -52,7 +52,7 @@ source_url: https://github.com/alphagov/govuk-developer-docs/blob/master/data/ap
 <% 
   unless application.private_repo?
     imported_docs = GitHubRepoFetcher.instance.docs(application.github_repo_name)
-    unless imported_docs.nil?
+    unless imported_docs.nil? || imported_docs.empty?
 %>
 ## Imported documents
 <%= partial(