--- a/source/templates/external_doc_template.html.erb
+++ b/source/templates/external_doc_template.html.erb
@@ -1,7 +1,8 @@
 ---
-layout: api_layout
+layout: application_layout
 parent: /apps.html
 hide_from_sitemap: true
+show_last_updated: true
 ---
 
 <%= ExternalDoc.parse(markdown, repository: repository, path: relative_path) %>