--- a/spec/app/github_repo_fetcher_spec.rb
+++ b/spec/app/github_repo_fetcher_spec.rb
@@ -213,4 +213,35 @@ RSpec.describe GitHubRepoFetcher do
       expect(instance.docs(repo_name)).to eq(nil)
     end
   end
+
+  describe "#nest_docs" do
+    it "returns a nested array of docs according to their original dir structure" do
+      top_level_doc = { title: "a top level doc", path: "/apps/some-app/foo.html" }
+      nested_doc = { title: "a nested doc", path: "/apps/some-app/subdir/bar.html" }
+      another_top_level_doc = { title: "another top level doc", path: "/apps/some-app/baz.html" }
+      another_nested_doc = { title: "another nested doc", path: "/apps/some-app/another-subdir/something.html" }
+      a_deeper_nested_doc = { title: "a deeper nested doc", path: "/apps/some-app/another-subdir/deep/labyrinth.html" }
+
+      flat_docs = [
+        top_level_doc,
+        nested_doc,
+        another_top_level_doc,
+        another_nested_doc,
+        a_deeper_nested_doc,
+      ]
+
+      expect(GitHubRepoFetcher.new.nest_docs(flat_docs)).to eq([
+        top_level_doc,
+        { title: "subdir/", docs: [nested_doc] },
+        another_top_level_doc,
+        {
+          title: "another-subdir/",
+          docs: [
+            another_nested_doc,
+            { title: "deep/", docs: [a_deeper_nested_doc] },
+          ],
+        },
+      ])
+    end
+  end
 end