Previously, we only imported top-level markdown files, of the form `docs/*.md`. Nested files - `docs/**/*.md` - were ignored. This is somewhat surprising behaviour and leads to an inconsistent mental model of how the dev docs importing works.

This was particularly problematic for apps like Smart Answers, which make heavy use of nested documentation: https://github.com/alphagov/smart-answers/tree/9e8366b4cb989fe2567d1fb3439261ee56749f39/docs

This PR now imports markdown files recursively. It also includes some necessary design changes as a result. Note: I did consider using the [Contents list component](https://components.publishing.service.gov.uk/component-guide/contents_list) for the design changes, but there was no way of creating non-clickable 'headings' for each subdirectory of links (other than passing `active: true` to the element, which would break accessibility).

Trello: https://trello.com/c/lJzIkFYP/218-import-subfolders-of-docs

## Performance impact

Importing more documents naturally has a performance impact. The new search.json is 9.1 MB in size, uncompressed. This is compared to 7.3 MB on production (which is 1.7 MB gzipped), so approximately a 20% increase.

The build time has increased too, from just over 3m 30s to between 4 and 5 minutes. This is an unavoidable consequence of making many more HTTP requests to the GitHub API. Locally, such requests are cached for an hour, so this has no noticeable effect on local development, but pull requests will be slower to receive an automated ✅. We should explore whether it's possible for GitHub Actions to make use of a shared cache, to have parity with the local dev experience. This is an optimisation we can look at in a future PR.

## Screenshots

These screenshots are from the [govuk-docker app page](https://docs.publishing.service.gov.uk/apps/govuk-docker.html):

### Imported documents list

| Before | After |
|--------|------|
| <img width="383" alt="Screenshot 2021-04-01 at 08 39 51" src="https://user-images.githubusercontent.com/5111927/113260377-1ba65e80-92c6-11eb-90e1-cb285512c8fb.png">| <img width="579" alt="Screenshot 2021-05-07 at 08 22 12" src="https://user-images.githubusercontent.com/5111927/117413010-5f1b6a80-af0d-11eb-8f6d-c2a32950e7f0.png"> |

### Breadcrumbs

This also shows the new `application_layout` after removal of `api_layout`:

> <img width="1127" alt="Screenshot 2021-03-31 at 16 50 04" src="https://user-images.githubusercontent.com/5111927/113173903-c6286e00-9241-11eb-8af8-e33e7cb912a1.png">