@benthorner 2021-04-27
Thanks for going over it again @ChrisBAshton, that's much simpler without the nested hash recursion ⛰️. I've now had a look at the whole PR, and have noticed one more major thing that's worth fixing. I also found a couple of things that you marked as resolved, but appear to be unchanged; I've made fresh comments for those, so it's clear.

Beyond that, I'm still interested in your thoughts on a couple of follow-up suggestions:

- I'm wondering if we could introduce a more useful layout for external docs, that shows the list of them for the repo

- Could we move the README into the list of imported docs?

Neither are blocking this PR, but it's a relevant opportunity to discuss them. I still haven't had a chance to try this out locally, but otherwise, I think it's pretty much ready to go ⚡.