@ChrisBAshton 2021-05-06
Thanks @benthorner - this should be ready for re-review now. As to your two other points:

> I'm wondering if we could introduce a more useful layout for external docs, that shows the list of them for the repo

Yes, I think you suggested a page like https://docs.publishing.service.gov.uk/manual.html but for external docs? Interesting idea, which we could consider for a follow-up PR, probably after the introduction of analytics where we can measure the usefulness of 'discovery' oriented pages like the Manual page.

> Could we move the README into the list of imported docs?

That's a good idea - someone actually reached out and said the [View Source doesn't match up with what they're seeing](https://twitter.com/kaihendry/status/1379298494616231943), so I think the page currently looks a bit like a README page when really the README is just one part of the page. Having it as an imported link instead would make that clearer. Again, may be worth considering splitting that out after we have analytics, so that we can measure the usefulness of having it on one page vs having it separate.