@benthorner 2021-05-06
I've had a quick chance to try this locally, and I'm seeing something strange for Content Store: the full path is being shown for nested docs without a title. Not a showstopper, so could be addressed in a follow-up PR.

<img width="540" alt="Screenshot 2021-05-06 at 12 20 37" src="https://user-images.githubusercontent.com/9029009/117290580-e0b7bd80-ae65-11eb-9015-f7e635de7013.png">

(I've got the latest version, but it's worth someone else corroborating this too.)