@ChrisBAshton 2021-05-06
> I've had a quick chance to try this locally, and I'm seeing something strange for Content Store: the full path is being shown for nested docs without a title.

I was seeing that, then I ran `rm -rf .cache` and ran the startup script again, where it worked fine 🤔