@benthorner 2021-05-06
> I was seeing that, then I ran rm -rf .cache and ran the startup script again, where it worked fine

So it does. Thanks for the pointer - still good to go then 👍.